Разбили проект на модули data.js и util.js точкой входа является файл main.js
