+++ package.json
@@ -9,6 +9,7 @@
   },
   "bugs": {
     "url": "https://github.com/htmlacademy-javascript/1708789-keksobooking-22/issues"
+    
   },
   "devDependencies": {
     "browser-sync": "2.26.14",
@@ -23,5 +24,6 @@
   },
   "engines": {
     "node": ">=12.0.0"
+    
   }
 }
